requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
google-genai==1.9.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0

